# Source/server/swagger.yaml
swagger: "2.0"
info:
  description: "API của Website Hỗ trợ đào tạo SGU"
  version: "1.0.0"
  title: "Website Hỗ trợ đào tạo SGU"
  contact:
    email: ""
host: ""
...
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
  /admin/chuc-nang/DanhSachChucNang:
    get: 
      tags:
        - "ChucNang"
      summary: Admin lấy danh sách chức năng
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        required: true
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        required: true
        description: giới hạn số lượng 1 trang, giá trị > 0
      - in: query
        name: keyword
        type: string
        required: true
        description: mã chức năng hoặc tên chức năng
      responses:
        200: 
          description: Lấy danh sách chức năng thành công
          schema:
            type: object
            example: {
              "success": true,
              "message": "Lấy danh sách chức năng thành công",
            }
        400:
          description: Lấy danh sách chức năng thất bại
          schema:
            type: object
            example: {
              "success": false,
              "message": "Lấy danh sách chức năng thất bại",
            }
      security:
      - ApiKeyAuth: []
  /admin/chuc-nang/ChiTietChucNang/{MaCN}:
    get:
      tags:
        - "ChucNang"
      summary: "Xem chi tiết thông tin chức năng"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaCN
          type: string
          required: true
          description: Mã chức năng
      responses:
        200:
          description: Xem chi tiết thông tin chức năng thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Xem chi tiết thông tin chức năng thành công.",
            }
        400:
          description: Xem chi tiết thông tin chức năng thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message" : "Xem chi tiết thông tin chức năng thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/chuc-nang/Them:
    post: 
      tags:
        - "ChucNang"
      summary: Thêm chức năng
      produces:
        - multipart/form-data    
      parameters:
        - in: formData
          name: MaCN
          type: string
          description: Mã chức năng
          required: true
        - in: formData
          name: TenChucNang
          type: string
          description: Tên chức năng
          required: true
        - in: formData
          name: Hinh
          type: file
          description: Hình icon chức năng
          required: true
      responses:
        200:
          description: Thêm chức năng thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Thêm chức năng thành công.",
            }          
        400:
          description: Thêm chức năng thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message": "Thêm chức năng thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/chuc-nang/ChinhSua/{MaCN}:
    put: 
      tags:
        - "ChucNang"
      summary: Chỉnh sửa thông tin chức năng
      produces:
        - multipart/form-data    
      parameters:
        - in: path
          name: MaCN
          type: string
          required: true
          description: Mã chức năng
        - in: formData
          name: TenChucNang
          type: string
          description: Tên chức năng
          required: true
        - in: formData
          name: Hinh
          type: file
          description: Hình icon chức năng
          required: true
      responses:
        200:
          description: Chỉnh sửa thông tin chức năng thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Chỉnh sửa thông tin chức năng thành công.",
            }          
        400:
          description: Chỉnh sửa thông tin chức năng thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message": "Chỉnh sửa thông tin chức năng thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/chuc-nang/Xoa/{MaCN}:
    delete:
      tags:
        - "ChucNang"
      summary: "Xóa thông tin chức năng"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaCN
          type: string
          required: true
          description: Mã chức năng
      responses:
        200:
          description: Xóa thông tin chức năng thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Xóa thông tin chức năng thành công.",
            }
        400:
          description: Xóa thông tin chức năng thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message" : "Xóa thông tin chức năng thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /admin/nganh/DanhSachNganh:
    get: 
      tags:
        - "Nganh"
      summary: Admin lấy danh sách ngành
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        required: true
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        required: true
        description: giới hạn số lượng 1 trang, giá trị > 0
      - in: query
        name: keyword
        type: string
        required: true
        description: mã ngành hoặc tên ngành
      responses:
        200: 
          description: Lấy danh sách ngành thành công
          schema:
            type: object
            example: {
              "success": true,
              "message": "Lấy danh sách ngành thành công",
            }
        400:
          description: Lấy danh sách ngành thất bại
          schema:
            type: object
            example: {
              "success": false,
              "message": "Lấy danh sách ngành thất bại",
            }
      security:
      - ApiKeyAuth: []
  /admin/nganh/ChiTietNganh/{MaNganh}:
    get:
      tags:
        - "Nganh"
      summary: "Chi tiết thông tin ngành"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaNganh
          type: string
          required: true
          description: Mã ngành
      responses:
        200:
          description: Chi tiết thông tin ngành thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Chi tiết thông tin ngành thành công.",
            }
        400:
          description: Chi tiết thông tin ngành thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message" : "Chi tiết thông tin ngành thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/nganh/Them:
    post: 
      tags:
        - "Nganh"
      summary: Thêm ngành
      produces:
        - application/json     
      parameters:
        - in: body
          name: MaNganh
          type: string
          required: true
          description: Mã ngành
        - in: body
          name: TenNganh
          type: string
          required: true
          description: Tên ngành
      responses:
        200:
          description: Thêm ngành thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Thêm ngành thành công.",
            }          
        400:
          description: Thêm ngành thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message": "Thêm ngành thất bại.",
            }
      security:
        - ApiKeyAuth: [] 
  /admin/nganh/ChinhSua/{MaNganh}:
    put: 
      tags:
        - "Nganh"
      summary: Chỉnh sửa ngành
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaNganh
          type: string
          required: true
          description: Mã ngành
        - in: body
          name: TenNganh
          type: string
          required: true
          description: Tên ngành
      responses:
        200:
          description: Chỉnh sửa ngành thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Chỉnh sửa ngành thành công.",
            }          
        400:
          description: Chỉnh sửa ngành thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message": "Chỉnh sửa ngành thất bại.",
            }
      security:
        - ApiKeyAuth: []     
  /admin/nganh/Xoa/{MaNganh}:
    delete:
      tags:
        - "Nganh"
      summary: "Xóa thông tin ngành"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaNganh
          type: string
          required: true
          description: Mã ngành
      responses:
        200:
          description: Xóa thông tin ngành thành công.
          schema:
            type: object
            example: {
              "success": true,
              "message": "Xóa thông tin ngành thành công.",
            }
        400:
          description: Xóa thông tin ngành thất bại.
          schema:
            type: object
            example: {
              "success": false,
              "message" : "Xóa thông tin ngành thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: headers
    name: x-access-token